refactor(utils): Move src/utils to src/lib/auth
