+++ src/lib/auth/password-validator.ts
