+++ src/lib/auth/tests/auth-token.test.ts
@@ -1,4 +1,4 @@
-import { generateAuthTokenAndSetCookie } from '@/utils/auth-token'
+import { generateAuthTokenAndSetCookie } from '@/lib/auth/auth-token'
 import type { Context } from 'hono'
 import { setSignedCookie } from 'hono/cookie'
 import { HTTPException } from 'hono/http-exception'
@@ -8,7 +8,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 // Mock dependencies
 vi.mock('hono/cookie')
 vi.mock('hono/jwt')
-vi.mock('@/lib/db')
+vi.mock('@/db')
 
 // Mock Context
 const mockContext = {
