+++ src/routes/auth/change-password.ts
@@ -1,11 +1,11 @@
 import { getUserPasswordByUserId, setUserPasswordByUserId } from '@/db'
+import { passwordValidator } from '@/lib/auth/password-validator'
 import {
   errorResponse,
   jsonContentRequired,
   jsonMessageContent,
 } from '@/lib/helper'
 import { authMiddleware, authMiddlewareSchema } from '@/middleware/auth'
-import { passwordValidator } from '@/utils/password-validator'
 import { createRoute, z } from '@hono/zod-openapi'
 import type { Context } from 'hono'
